--- service/src/main/java/bio/terra/profile/app/common/MDCHandlingException.java
@@ -1,27 +1,0 @@
-package bio.terra.profile.app.common;
-
-import bio.terra.common.exception.InternalServerErrorException;
-import java.util.List;
-
-public class MDCHandlingException extends InternalServerErrorException {
-
-  public MDCHandlingException(String message) {
-    super(message);
-  }
-
-  public MDCHandlingException(String message, Throwable cause) {
-    super(message, cause);
-  }
-
-  public MDCHandlingException(Throwable cause) {
-    super(cause);
-  }
-
-  public MDCHandlingException(String message, List<String> causes) {
-    super(message, causes);
-  }
-
-  public MDCHandlingException(String message, Throwable cause, List<String> causes) {
-    super(message, cause, causes);
-  }
-}
